Rename directory from 'models' to 'entities'
