+++ src/entities/Conference.ts
