+++ src/entities/Config.ts
