+++ src/features/request/RequestList.tsx
@@ -23,7 +23,7 @@
 //
 
 import { IRDETypes } from "../../utils/IRDE";
-import { Request } from "../../models/Request";
+import { Request } from "../../entities/Request";
 import { RequestListIRDE } from "./RequestViewModel";
 import { Box, CircularProgress, Grid, Typography } from "@mui/material";
 import { RequestCard } from "./RequestCard";
