+++ src/features/session/SessionCard.tsx
@@ -26,7 +26,7 @@ import { IdAndName, SessionItem } from "./SessionViewModel";
 import { Avatar, Card, CardActionArea, Grid, styled, Typography } from "@mui/material";
 import { Check } from "@mui/icons-material";
 import { CaseInsensitiveSearch } from "../../utils/CaseInsensitiveSearch";
-import { Speaker } from "../../models/Session";
+import { Speaker } from "../../entities/Session";
 
 export interface SessionCardProps {
   sessionItem: SessionItem;
